requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.36.0
pydantic==2.11.9
PyJWT==2.10.1
python-dotenv==1.1.1
